+++ CHANGELOG.md
@@ -1,3 +1,32 @@
+# [1.11.0-dev.3](https://github.com/crimera/piko/compare/v1.11.0-dev.2...v1.11.0-dev.3) (2024-03-08)
+
+
+### Bug Fixes
+
+* **Twitter:** Download patch ([f87dee7](https://github.com/crimera/piko/commit/f87dee798632528595ac3efab1349d7a3fecafa0))
+
+# [1.11.0-dev.2](https://github.com/crimera/piko/compare/v1.11.0-dev.1...v1.11.0-dev.2) (2024-03-04)
+
+
+### Features
+
+* **Twitter:** Hide FAB ([4cdf50f](https://github.com/crimera/piko/commit/4cdf50fb036ed8596c9e4a3e3f1702ae78525d28))
+* **Twitter:** Hide FAB Menu Buttons ([4f25bd6](https://github.com/crimera/piko/commit/4f25bd60888b70ccd4518a42459583b6ef8fcc54))
+* **Twitter:** Hide Promoted Trends ([1be736e](https://github.com/crimera/piko/commit/1be736e31ae11d21a62f7a3f4aad2689647ada92))
+
+# [1.11.0-dev.1](https://github.com/crimera/piko/compare/v1.10.0...v1.11.0-dev.1) (2024-03-03)
+
+
+### Bug Fixes
+
+* unshort url for 10.30 ([ef3e00f](https://github.com/crimera/piko/commit/ef3e00f80fc469e5f09f8d7e43136791dbd94872))
+
+
+### Features
+
+* no t.co links ([776e700](https://github.com/crimera/piko/commit/776e7007b571cc2803d566891c72a42057297bb6))
+* **Twitter:** Disable in favor of official patch bundle ([9a81fab](https://github.com/crimera/piko/commit/9a81fabed7457d1be967250faa548a67accc27bc))
+
 # [1.10.0](https://github.com/crimera/piko/compare/v1.9.1...v1.10.0) (2024-03-03)